+++ src/main/java/de/chrgroth/jsonstore/store/AbstractJsonStore.java
@@ -59,20 +59,21 @@ public abstract class AbstractJsonStore<T, P> {
     protected final boolean deepSerialize;
     protected final Map<Integer, VersionMigrationHandler> migrationHandlers;
 
-    protected AbstractJsonStore(Class<T> payloadClass, Integer payloadTypeVersion, boolean singleton, FlexjsonHelper flexjsonHelper, File storage, Charset charset,
+    protected AbstractJsonStore(String uid, Class<T> payloadClass, Integer payloadTypeVersion, boolean singleton, FlexjsonHelper flexjsonHelper, File storage, Charset charset,
             boolean prettyPrint, boolean autoSave, boolean deepSerialize, VersionMigrationHandler... migrationHandlers) {
-        this(payloadClass, payloadTypeVersion, singleton, flexjsonHelper, storage, charset, "", prettyPrint, autoSave, deepSerialize, migrationHandlers);
+        this(uid, payloadClass, payloadTypeVersion, singleton, flexjsonHelper, storage, charset, "", prettyPrint, autoSave, deepSerialize, migrationHandlers);
     }
 
-    protected AbstractJsonStore(Class<T> payloadClass, Integer payloadTypeVersion, boolean singleton, FlexjsonHelper flexjsonHelper, File storage, Charset charset,
+    protected AbstractJsonStore(String uid, Class<T> payloadClass, Integer payloadTypeVersion, boolean singleton, FlexjsonHelper flexjsonHelper, File storage, Charset charset,
             String fileNameExtraPrefix, boolean prettyPrint, boolean autoSave, boolean deepSerialize, VersionMigrationHandler... migrationHandlers) {
         this.flexjsonHelper = flexjsonHelper;
         metadata = new JsonStoreMetadata<>();
+        metadata.setUid(uid);
         metadata.setPayloadType(payloadClass.getName());
         metadata.setPayloadTypeVersion(payloadTypeVersion);
         metadata.setSingleton(singleton);
         metadata.setCreated(new Date());
-        this.file = storage != null ? new File(storage, FILE_PREFIX + FILE_SEPARATOR + fileNameExtraPrefix + payloadClass.getName() + FILE_SEPARATOR + FILE_SUFFIX) : null;
+        this.file = storage != null ? new File(storage, FILE_PREFIX + FILE_SEPARATOR + fileNameExtraPrefix + uid + FILE_SEPARATOR + FILE_SUFFIX) : null;
         this.charset = charset;
         this.prettyPrint = prettyPrint;
         this.autoSave = autoSave;
@@ -105,7 +106,16 @@ public abstract class AbstractJsonStore<T, P> {
             fileSize = FileUtils.sizeOf(file);
         }
 
-        return new JsonStoreMetrics(metadata.getPayloadType(), size(), metadata.getModified(), fileSize);
+        return new JsonStoreMetrics(metadata.getUid(), metadata.getPayloadType(), size(), metadata.getModified(), fileSize);
+    }
+
+    /**
+     * Returns the stores uid.
+     *
+     * @return uid
+     */
+    public String getUid() {
+        return metadata.getUid();
     }
 
     /**
@@ -149,7 +159,7 @@ public abstract class AbstractJsonStore<T, P> {
                 Files.write(file.toPath(), Arrays.asList(json), charset, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING, StandardOpenOption.WRITE);
             }
             stopwatch.stop();
-            LOG.info(metadata.getPayloadType() + ": saving json to file took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
+            LOG.info(metadata.getUid() + ": saving json to file took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
         } catch (IOException e) {
             LOG.error("Unable to write file content, skipping file during store: " + file.getAbsolutePath() + "!!", e);
         }
@@ -182,7 +192,7 @@ public abstract class AbstractJsonStore<T, P> {
             return deepSerialize ? serializer.deepSerialize(metadata) : serializer.serialize(metadata);
         } finally {
             stopwatch.stop();
-            LOG.info(metadata.getPayloadType() + ": converting to json took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
+            LOG.info(metadata.getUid() + ": converting to json took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
         }
     }
 
@@ -210,7 +220,7 @@ public abstract class AbstractJsonStore<T, P> {
                 json = Files.lines(file.toPath(), charset).parallel().filter(line -> line != null && !"".equals(line.trim())).map(String::trim).collect(Collectors.joining());
             }
             stopwatch.stop();
-            LOG.info(metadata.getPayloadType() + ": loading json from file took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
+            LOG.info(metadata.getUid() + ": loading json from file took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
         } catch (Exception e) {
             throw new JsonStoreException("Unable to read file content: " + file.getAbsolutePath() + "!!", e);
         }
@@ -242,7 +252,7 @@ public abstract class AbstractJsonStore<T, P> {
         Stopwatch stopwatch = Stopwatch.createStarted();
         Object genericStructureRaw = new JSONTokener(json).nextValue();
         stopwatch.stop();
-        LOG.info(metadata.getPayloadType() + ": raw parsing from json took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
+        LOG.info(metadata.getUid() + ": raw parsing from json took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
         if (genericStructureRaw == null) {
             return;
         }
@@ -302,7 +312,7 @@ public abstract class AbstractJsonStore<T, P> {
                 }
 
                 // invoke handler per instance, so you don't have to deal with wrapping outer list by yourself
-                LOG.info(metadata.getPayloadType() + ": migrating to version " + i + " using " + migrationHandler);
+                LOG.info(metadata.getUid() + ": migrating to version " + i + " using " + migrationHandler);
                 try {
                     Stopwatch stopwatch = Stopwatch.createStarted();
                     if (isSingleton) {
@@ -314,9 +324,9 @@ public abstract class AbstractJsonStore<T, P> {
                     }
                     stopwatch.stop();
                     migrated = true;
-                    LOG.info(metadata.getPayloadType() + ": migrating to version " + i + " took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
+                    LOG.info(metadata.getUid() + ": migrating to version " + i + " took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
                 } catch (Exception e) {
-                    throw new JsonStoreException("failed to migrate " + metadata.getPayloadType() + " from version " + i + " to " + (i + 1) + ": " + e.getMessage() + "!!", e);
+                    throw new JsonStoreException("failed to migrate " + metadata.getUid() + " from version " + i + " to " + (i + 1) + ": " + e.getMessage() + "!!", e);
                 }
             }
         }
@@ -354,7 +364,7 @@ public abstract class AbstractJsonStore<T, P> {
                 metadata.setModified(now);
             }
             stopwatch.stop();
-            LOG.info(metadata.getPayloadType() + ": deserializing from json took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
+            LOG.info(metadata.getUid() + ": deserializing from json took " + stopwatch.elapsed(TimeUnit.MILLISECONDS) + "ms");
 
             // metadata refresh callback
             metadataRefreshed();
@@ -374,7 +384,7 @@ public abstract class AbstractJsonStore<T, P> {
     public final void drop() {
         if (isPersistent()) {
             try {
-                LOG.info(metadata.getPayloadType() + ": dropping strage file");
+                LOG.info(metadata.getUid() + ": dropping strage file");
                 synchronized (file) {
                     Files.deleteIfExists(file.toPath());
                 }